Edit Component to Login Template (Start)
